Added a mainWindow and added an 'open file' function.
Added a mainWindow class which contains the quiz window.  The mainwindow includes menus and options to open a dictionary file, quit, and show information about the project and about QT.  The quiz widget is better rounded out - the user can now press enter to check a word, and text in the answer box is highlighted to make it easier for the user to keep typing without using tab to navigate buttons.  All interface elements are cross-platform, isolated fromt the backend, and have alt-keys and shortcuts attached to them.