Added a new framework for storing lists and users
Changed the method of storing lists to be focused on connections, rather
than std::strings in the Dictionary class.  The Dictionary class is now
hardly used: lists of Connections are loaded into a QuizList, which
takes care of ordering words to be provided to the VocabQuiz object.

VocabQuiz may need to be removed soon as it has become something of a
wrapper class for QuizList, but it may be a nice layer of abstraction.

UserProfiles have been added, with a view of keeping track of statistics
on each word for each user.