Added simple profile saving and loading.
Saving and loading profiles mostly works, but is buggy still and certaintly not as robust as it eventually should be. A major problem is that languages not sorted alphabetically will get out of sync with their words so that the languages are switched in order, but the words are not. Other problems are simply that the loader will crash if the format is not as expected.

This will only work on Linux right now. This commit is made to move the code to windows to get the code working on windows.